@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace spec\Omniva;
 
 use PhpSpec\ObjectBehavior;
@@ -8,12 +10,12 @@ use Omniva\PickupPoint;
 
 class AddressSpec extends ObjectBehavior
 {
-    function it_is_initializable()
+    function it_is_initializable(): void
     {
         $this->shouldHaveType(Address::class);
     }
 
-    public function it_should_have_name()
+    public function it_should_have_name(): void
     {
         $name = 'name surname';
 
@@ -21,7 +23,7 @@ class AddressSpec extends ObjectBehavior
         $this->getName()->shouldReturn($name);
     }
 
-    public function it_should_have_phone()
+    public function it_should_have_phone(): void
     {
         $mobile = '+37060000000';
 
@@ -29,7 +31,7 @@ class AddressSpec extends ObjectBehavior
         $this->getPhone()->shouldReturn($mobile);
     }
 
-    public function it_should_have_email()
+    public function it_should_have_email(): void
     {
         $email = 'ab@nebijokit.lt';
 
@@ -39,7 +41,7 @@ class AddressSpec extends ObjectBehavior
         $this->getEmail()->shouldReturn($email);
     }
 
-    public function it_should_have_country_code()
+    public function it_should_have_country_code(): void
     {
         $code = 'LT';
 
@@ -47,23 +49,23 @@ class AddressSpec extends ObjectBehavior
         $this->getCountryCode()->shouldReturn($code);
     }
 
-    public function it_should_trigger_deprecation_when_setting_terminal()
+    public function it_should_trigger_deprecation_when_setting_terminal(): void
     {
         $identifier = '88822';
         $this->shouldTrigger(E_USER_DEPRECATED)->duringSetTerminal($identifier);
     }
 
-    public function it_should_trigger_deprecation_when_getting_terminal()
+    public function it_should_trigger_deprecation_when_getting_terminal(): void
     {
         $this->shouldTrigger(E_USER_DEPRECATED)->duringGetTerminal();
     }
 
-    public function it_should_trigger_deprecation_when_checking_whether_terminal_is_set()
+    public function it_should_trigger_deprecation_when_checking_whether_terminal_is_set(): void
     {
         $this->shouldTrigger(E_USER_DEPRECATED)->duringHasTerminal();
     }
 
-    public function it_should_have_city()
+    public function it_should_have_city(): void
     {
         $city = 'Vilnius';
 
@@ -71,7 +73,7 @@ class AddressSpec extends ObjectBehavior
         $this->getCity()->shouldReturn($city);
     }
 
-    public function it_should_have_street()
+    public function it_should_have_street(): void
     {
         $street = 'Konstitucijos pr. 8';
 
@@ -79,7 +81,7 @@ class AddressSpec extends ObjectBehavior
         $this->getStreet()->shouldReturn($street);
     }
 
-    public function it_should_have_postcode()
+    public function it_should_have_postcode(): void
     {
         $code = '00000';
 
@@ -87,7 +89,7 @@ class AddressSpec extends ObjectBehavior
         $this->getPostCode()->shouldReturn($code);
     }
 
-    public function it_should_allow_to_set_pickup_point(PickupPoint $point)
+    public function it_should_allow_to_set_pickup_point(PickupPoint $point): void
     {
         $this->setPickupPoint($point)->shouldHaveType(Address::class);
         $this->getPickupPoint()->shouldReturn($point);
