@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace spec\Omniva;
 
 use Omniva\Client;
@@ -8,17 +10,17 @@ use GuzzleHttp\Client as HttpClient;
 
 class ClientSpec extends ObjectBehavior
 {
-    public function let()
+    public function let(): void
     {
         $this->beConstructedWith('username', 'password');
     }
 
-    function it_is_initializable()
+    function it_is_initializable(): void
     {
         $this->shouldHaveType(Client::class);
     }
 
-    public function it_should_return_http_client()
+    public function it_should_return_http_client(): void
     {
         $this->getHttpClient()->shouldHaveType(HttpClient::class);
     }
