@@ -1,28 +1,29 @@
 <?php
 
+declare(strict_types=1);
+
 namespace spec\Omniva;
 
 use Omniva\Parcel;
 use Omniva\Service;
 use Omniva\Address;
 use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
 
 class ParcelSpec extends ObjectBehavior
 {
-    function it_is_initializable()
+    function it_is_initializable(): void
     {
         $this->shouldHaveType(Parcel::class);
     }
 
-    public function it_should_allow_to_set_weight()
+    public function it_should_allow_to_set_weight(): void
     {
         $weight = 120.0;
         $this->setWeight($weight)->shouldHaveType(Parcel::class);
         $this->getWeight()->shouldReturn($weight);
     }
 
-    public function it_should_allow_to_set_comment()
+    public function it_should_allow_to_set_comment(): void
     {
         $comment = 'comment';
         $this->hasComment()->shouldReturn(false);
@@ -31,14 +32,14 @@ class ParcelSpec extends ObjectBehavior
         $this->getComment()->shouldReturn($comment);
     }
 
-    public function it_should_allow_to_set_foreign_id()
+    public function it_should_allow_to_set_foreign_id(): void
     {
         $foreignId = 'order-1';
         $this->setPartnerId($foreignId)->shouldHaveType(Parcel::class);
         $this->getPartnerId()->shouldReturn($foreignId);
     }
 
-    public function it_should_return_cod_amount()
+    public function it_should_return_cod_amount(): void
     {
         $amount = 12.22;
         $this->getCodAmount()->shouldReturn(null);
@@ -46,17 +47,17 @@ class ParcelSpec extends ObjectBehavior
         $this->getCodAmount()->shouldReturn($amount);
     }
 
-    public function it_should_have_bank_account()
+    public function it_should_have_bank_account(): void
     {
         $bank = 'LT000121';
         $this->setBankAccount($bank)->shouldHaveType(Parcel::class);
         $this->getBankAccount()->shouldReturn($bank);
     }
 
-    public function it_should_be_possible_to_add_service()
+    public function it_should_be_possible_to_add_service(): void
     {
-        $sms = Service::SMS();
-        $cod = Service::COD();
+        $sms = Service::SMS;
+        $cod = Service::COD;
 
         $this->hasServices()->shouldReturn(false);
         $this->addService($sms)->shouldHaveType(Parcel::class);
@@ -68,7 +69,7 @@ class ParcelSpec extends ObjectBehavior
         $this->getServices()->count()->shouldReturn(2);
     }
 
-    public function it_should_have_address()
+    public function it_should_have_address(): void
     {
         $address = new Address();
         $address
@@ -90,7 +91,7 @@ class ParcelSpec extends ObjectBehavior
         $this->getReturnee()->shouldReturn($address);
     }
 
-    public function it_should_have_tracking_nubmer()
+    public function it_should_have_tracking_nubmer(): void
     {
         $number = 'AB0009213434LT';
         $this->hasTrackingNumber()->shouldReturn(false);
